wavelength: explicitly perform calculation as float64
As there is no longer an implicit promotion from float32 to float64 in
numpy 2.x, we need to explicitly ensure we have sufficient numerical
precision.

+++ src/ptychography40/reconstruction/common.py
@@ -39,7 +39,7 @@ def wavelength(U):
     c = const.speed_of_light  # Speed of light
     m_0 = const.electron_mass  # Electron rest mass
 
-    T = e*U*1000
+    T = e*np.float64(U)*1000
     lambda_e = h*c/(math.sqrt(T**2+2*T*m_0*(c**2)))
     return lambda_e
